Fix failing junit tests
