+++ web3/src/main/java/org/hiero/mirror/web3/state/core/MapReadableKVState.java
@@ -2,15 +2,12 @@
 
 package org.hiero.mirror.web3.state.core;
 
-import com.google.common.collect.ForwardingConcurrentMap;
 import com.swirlds.state.spi.ReadableKVState;
 import com.swirlds.state.spi.ReadableKVStateBase;
 import jakarta.annotation.Nonnull;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.ConcurrentMap;
-import org.hiero.mirror.web3.common.ContractCallContext;
 
 /**
  * A simple implementation of {@link ReadableKVState} backed by a
@@ -35,12 +32,7 @@ public class MapReadableKVState<K, V> extends ReadableKVStateBase<K, V> {
      * @param backingStore The backing store to use
      */
     public MapReadableKVState(@Nonnull final String stateKey, @Nonnull final Map<K, V> backingStore) {
-        super(stateKey, new ForwardingConcurrentMap<K, V>() {
-            @Override
-            protected ConcurrentMap<K, V> delegate() {
-                return ContractCallContext.get().getReadCacheState(stateKey);
-            }
-        });
+        super(stateKey, new ReadableCachedForwardingConcurrentMap<>(stateKey));
         this.backingStore = Objects.requireNonNull(backingStore);
     }
